Fixed mistake in try catch
